lint test files
