+++ test/postList_spec.js
@@ -1,87 +1,92 @@
-import { expect } from 'chai';
-import * as types from '../src/constants';
-import reducer from '../src/routes/PostList/reducer';
+import { expect } from 'chai'
+import * as types from '../src/constants'
+import reducer from '../src/routes/Post/reducer'
 
 // Remove this
-import fakeDB from '../src/server/fakeDB.js';
+import fakeDB from '../src/server/fakeDB.js'
 
-describe('PostList Reducer', () => {
+describe('Post Reducer', () => {
 
   it('should return default state if action is undefined', () => {
-    const initialState = [];
-    const nextState = reducer(initialState, 'BLAH');
-    expect(nextState).to.deep.equal(initialState);
-  });
+    const initialState = []
+    const nextState = reducer(initialState, 'BLAH')
+    expect(nextState).to.deep.equal(initialState)
+  })
 
   it('should handle LOAD_POSTS_REQUEST', () => {
     const initialState = {
       lastFetched: null,
       isLoading: false,
       error: null,
-      data: [],
-    };
+      data: []
+    }
 
     const action = {
-      type: types.LOAD_POSTS_REQUEST,
-    };
+      type: types.LOAD_POST_REQUEST
+    }
 
-    const nextState = reducer(initialState, action);
+    const nextState = reducer(initialState, action)
     expect(nextState).to.deep.equal({
       lastFetched: null,
       isLoading: true,
       error: null,
-      data: [],
-    });
-  });
+      data: []
+    })
+  })
 
   it('should handle LOAD_POSTS_SUCCESS', () => {
     const initialState = {
       lastFetched: null,
       isLoading: false,
       error: null,
-      data: [],
-    };
+      data: []
+    }
 
-    const currentTime = Date.now();
+    const post = {
+      id: '128sd043hd',
+      title: 'Cloth Talk Part I',
+      slug: 'cloth-talk-part-i',
+      content: 'Khaled Ipsum is a major key to success.'
+    }
+
+    const currentTime = Date.now()
+
     const action = {
-      type: types.LOAD_POSTS_SUCCESS,
-      payload: fakeDB,
+      type: types.LOAD_POST_SUCCESS,
+      payload: post,
       meta: {
-        lastFetched: currentTime,
+        lastFetched: currentTime
       }
-    };
+    }
 
-    const nextState = reducer(initialState, action);
+    const nextState = reducer(initialState, action)
     expect(nextState).to.deep.equal({
       lastFetched: currentTime,
       isLoading: false,
       error: null,
-      data: fakeDB,
-    });
-  });
+      data: post
+    })
+  })
 
-  it('should handle LOAD_POSTS_FAILURE', () => {
+  it('should handle LOAD_POST_FAILURE', () => {
     const initialState = {
       lastFetched: null,
       isLoading: false,
       error: null,
-      data: [],
-    };
-
+      data: {}
+    }
     const error = new Error('Invalid request')
-
     const action = {
-      type: types.LOAD_POSTS_FAILURE,
+      type: types.LOAD_POST_FAILURE,
       payload: error,
       error: true
-    };
-
-    const nextState = reducer(initialState, action);
+    }
+    const nextState = reducer(initialState, action)
     expect(nextState).to.deep.equal({
       lastFetched: null,
       isLoading: false,
-      error,
-      data: [],
-    });
-  });
-});
+      error: error,
+      data: {}
+    })
+  })
+})
